sample_content/GA Lab Examples/testing-rest-cat-api-views-lab-main/requirements.txt
asgiref==3.7.2
certifi==2024.12.14
distlib==0.3.9
Django==4.2.18
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.12.2
virtualenv==20.29.0
